ML-wastewise/FastAPI/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.22.0
tensorflow==2.18.0
numpy==1.26.4
python-dotenv==1.1.0
protobuf>=3.20.3,<6.0.0
google-generativeai==0.8.5
python-multipart==0.0.9
pillow==11.1.0
